pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pbdata"
version = "0.1.0"
description = "Protein binding dataset platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0.1",
  "typer>=0.12",
  "rich>=13.7",
  "gemmi",
  "pandas>=2.0",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.5",
]

[project.scripts]
pbdata = "pbdata.cli:app"
pbdata-gui = "pbdata.gui:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-p no:tmpdir"
markers = [
  "integration: tests that make real network calls",
]

[tool.ruff]
line-length = 100
target-version = "py311"
